Updated
List<Expense>: We use the List interface to create a flexible list that can store multiple Expense objects.
ArrayList<Expense>: The ArrayList is an implementation of the List interface that dynamically resizes as you add new elements.
The ArrayList already implements the List interface, which is part of the Java collections framework.